api/requirements.txt
# AIRecruiter v2 API Dependencies

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
pyodbc>=5.0.0
alembic>=1.13.0

# Settings & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP Client
httpx>=0.26.0

# Logging
structlog>=24.1.0

# Utilities
python-multipart>=0.0.6
pyhumps>=3.8.0

# AWS
boto3>=1.34.0

# AI
anthropic>=0.18.0

# PDF Processing
pypdf2>=3.0.0
python-docx>=1.1.0
weasyprint>=60.0

# SOAP Client (Workday)
zeep>=4.2.0

# Encryption
cryptography>=42.0.0
